Changed it to load migrations with ActiveRecord::Migrator.migrate('db/migrate') instead of duplicating them inside of spec/connection_and_schema.rb.
Tried some more to get controller tests to pass and overcame some errors
but am still getting 15 failures, most with this error:
     undefined method `authenticate' for nil:NilClass

At least the model specs are passing. That's at least worth something.
(ruby -S bundle exec rspec spec/models/page_spec.rb)

I've had to duplicate an alarming amount of initialization code from a
rails app in spec_helper.rb and it's clearly getting out of control! It
would probably be better to create an actual, complete rails app within
the spec directory and initialize that rather than trying to cherry pick
the bits and pieces that are necessary and hope that we initialize
everything in the correct order. It appears that devise does something
like this (in the test/rails_app directory), for example.