k3/inline_editor.js: After the content is saved, made it do a GET to retrieve the current state of the object from the response, and then try to update the page with this updated state data.
This is useful in case any other attributes of the resource have changed
as a result of the save you just did (perhaps updating one attribute of
your model automatically causes another one to be changed) or as a
result of someone else editing the record at the same time as you.  It
also solves the problem of the same field appearing on the page in
multiple places.

inline_editable helper: Added 'data-object-id' attribute. Added options argument.

Refactored the unbinding/binding of event handlers out to
bindEventHandlers so that we can call it from both the constructor and
mergeOptions (which gets used if we call inlineEditor() on an element
that already has an InlineEditor).

Removed dependency on rest_in_place, since wo probably won't be using
any other features/form-types from it, and copied what we were using
into inline_editor/core/inline_editor.js.