It was redundant to both set the class to 'left' and add an inline style 'float: left' so changed it to only set the class.
Renamed the class from 'left' to 'float_left' because it seems quite
possible that a site/theme might want a class named 'left' to *not* be
floated left and we don't want to conflict with them.

Added the styles for these classes to the k3/inline_editor.css
stylesheet so that users can get the behavior expected from the editor
(setting float left makes it float left) out of the box.