Made sure k3cms.authorization.load_authorization_files initializer is loaded *before* all the require_decorators initializers, because require_decorators causes User to be loaded, which includes RealUser module, which causes the permission sets to be loaded and cached in a class variable:
      def self.included(base)
        auth_sets = Rails.application.railties.engines.collect {|e| e.class.authorization.default_suggested_permission_set}
        @@k3cms_real_perms = auth_sets.compact.inject([].to_set) {|result,set| result + set.user + set.manager + set.admin}
      end

So if we load_authorization_files *after* that occurs, it is too late
because the (empty) list of permissions has already been loaded and
cached.

This is only a problem in production. In development, it gets reloaded
after every request so at least by the 2nd request, all the
authorization files will have been loaded.