Renamed :inside_head hook to just :head.
Made it so hook(:hook_name) also includes content_for(:hook_name) into
your layout, so that you don't have to have a separate
content_for(:head) in addition to your hook(:head). That seemed kind of
redundant. It the insertion points marked by hook() it seems likely that
we might also want a content_for() hook so that views have the ability
to add things to the layout at those points just as easily as railties!